Change generate_cc.bzl to avoid using colon when unnecessary. (#140)
generate_cc.bzl creates rules that call protoc with options like
--cpp_out=:some/path.  This is the form that protoc expects to take, and
it is usually fine.  However, on Windows, Bazel runs commands through an
MSYS shell, and when executing Windows-native commands, MSYS tries to
translate anything that looks POSIXy to be Windowsy.  Unfortunately,
here its heuristics kick in, thinking it looks like a POSIXy
colon-delimited PATH, and converts the colon to a semicolon, which is
then misinterpreted by protoc.

When there are no options, we omit the colon, which avoids this error
when building on Windows.  This still doesn't solve the case when
options do need to be passed, but fortunately that case has not yet come
up, so this is an acceptable stopgap solution.