Update to rules_go 0.7.0+
Adds a check in the require function to avoid strictness
if the pre-existing rule does not have a sha value set
(where strict is defined as "fail if we try to re-load / override
an external dependency where the sha256 values for the
rule don't match").

Removes 0.5.x from travis matrix and removes residual pip
dependency.