Allow overriding implicit cpp repositories options in other language rules.
This shows up when using a different version of google test to the one
protobuf expects to be built with.

An alternative is to explicitly call `cpp_proto_repositories()`, passing
any necessary arguments first (`overrides` in this case) and then call eg.
`python_proto_repositories(omit_cpp_repositories=True)`. This is less
obvious without reading the bazel rules themselves.

For reference, the error was:

    An existing http_archive rule 'com_google_googletest' was already loaded with a sha256 value of '...'.  Refusing to overwrite this with the requested value ('f87029f647276734ef076785f76652347993b6d13ac1cbb2d2e976e16d2f8137').
    Either remove the pre-existing rule from your WORKSPACE or exclude it from loading by rules_protobuf.