Fixing proto compiles from WORKSPACE root
_get_relative_directory blindly removes the first directory of
the proto file path in some cases. This is fine when it is removing
the "." returned by dirname when the BUILD file and the proto file
reside in the same directory, but in cases where the proto file is
in a subdirectory it causes the first subdirectory to be stripped.

fixes #201